Write a search heuristic called ModRestrictedForwardSelection.
The problem with RestrictedForwardSelection is that the algorithm is not
always able to return the number of attributes that the user desires. This
does not always affect classification accuracy negatively, but it may still
be inconvienient for a user.

ModRestrictedForwardSelection is my own version of RestrictedForwardSelection
which overcomes this problem. Instead of dividing the set size by M/i where
i is the iteration number, I divide the set size by a fixed divFactor which
can be calculated by taking the mth root of M (where m is the number of
attributes to select and M is the total number of attributes). A further
optimization is performed where remaining attributes from the old list are
automatically added to the best set if the size of the new list is less than
or equal to the number of remaining attributes required.

The committed code includes debug messages. I am also yet to write the
toString() method or any technical information and help messages.