Fix corner case bug in my ModRestrictedForwardSelection algorithm.
When selecting 2 attributes from a file containing 2 attributes, it was
getting a null pointer exception. The invariant that says that the
early return case would trigger, otherwise it would have had to trigger in the
previous step does not hold for the very first step. Thus, it could get a case
where the expected list size was greater than the number of attributes left to
fill it. By adding a check, I satisfy the base case for my induction.